refactor: Station 클래스에 있는 String name 변수를 StationName 클래스로 변환
